Feat: added repository

+++ src/main/java/org/example/web/controllers/BookShelfController.java
@@ -7,10 +7,13 @@ import org.example.web.dto.BookIdToRemove;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import javax.validation.Valid;
 
 @Controller
 @RequestMapping("/books")
@@ -34,21 +37,34 @@ public class BookShelfController {
     }
 
     @PostMapping("/save")
-    public String saveBook(Book book) {
-        bookService.save(book);
-        logger.info("Book saved. Current repository size =" + bookService.getAllBooks().size());
-        return "redirect:/books/shelf";
+    public String saveBook(@Valid Book book, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("book", book);
+            model.addAttribute("bookIdToRemove", new BookIdToRemove());
+            model.addAttribute("bookList", bookService.getAllBooks());
+            return "book_shelf";
+        } else {
+            bookService.save(book);
+            logger.info("Book saved. Current repository size =" + bookService.getAllBooks().size());
+            return "redirect:/books/shelf";
+        }
     }
 
     @PostMapping("/remove")
-    public String removeBook(BookIdToRemove bookIdToRemove) {
-        bookService.removeBookId(bookIdToRemove.getId());
-        logger.info("Book deleted by Id. Current repository size=" + bookService.getAllBooks().size());
-        return "redirect:/books/shelf";
+    public String removeBook(@Valid BookIdToRemove bookIdToRemove, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("book", new Book());
+            model.addAttribute("bookList", bookService.getAllBooks());
+            return "book_shelf";
+        } else {
+            bookService.removeBookId(bookIdToRemove.getId());
+            logger.info("Book deleted by Id. Current repository size=" + bookService.getAllBooks().size());
+            return "redirect:/books/shelf";
+        }
     }
 
     @PostMapping("/removeByRegex")
-    public String removeBookByRegex (@RequestParam(value = "queryRegex") String regex) {
+    public String removeBookByRegex(@RequestParam(value = "queryRegex") String regex) {
         bookService.removeBooksByRegex(regex);
         logger.info("Book deleted by regex Current repository size=" + bookService.getAllBooks().size());
         return "redirect:/books/shelf";
